requirements.txt
Django>=5.0,<6.0
python-decouple>=3.8
pillow>=10.0
psycopg2-binary>=2.9
django-htmx>=1.17
django-widget-tweaks>=1.5
django-crispy-forms>=2.1
crispy-tailwind>=1.0
django-extensions>=3.2
whitenoise>=6.6
django-ratelimit>=4.1

# Testing
pytest>=9.0
pytest-django>=4.11
pytest-cov>=7.0
faker>=39.0
